Better way around subscription-manager container detection
The container images are supposed to behave as "real systems" (rather
than containers). Because of historical/business reasons,
subscription-manager tries to detect whether it is run within a
container, and disables itself in that case. Before RHEL 9, this is done
only by checking for /etc/rhsm-host, which is a symlink pointing to
/run/secrets/rhsm where a docker plugin, or podman configuration, uses
to mount the subscription-manager bits from the host. In RHEL 9,
subscription-manager also checks for the well-known files of docker or
podman as part of this container detection.

The ways done to avoid this situation so far have been the following:
- hot-patch the subscription-manager to not check /etc/rhsm-host at all;
  this is super-unsupported in any way and form, and prone to break at
  any time without any pre-advice
- in RHEL 9, subscription-manager uses an internal environment variable
  called "SMDEV_CONTAINER_OFF" to disable the container detection: this
  variable has always been meant to be internal for development, and
  unfortunately it got widespread with no feedback to the
  subscription-manager team...

Hence, simplify the situation a bit:
- remove the hot-patching from SUSE images: those container images
  definitely do not ship with /etc/rhsm-host, so there is nothing to fix
- replace the hot-patching with the removal of the /etc/rhsm-host &
  /etc/pki/entitlement-host symlinks, so preventing subscription-manager
  to notice bits for it mounted in /run/secrets/rhsm
- keep the "SMDEV_CONTAINER_OFF" hack in RHEL 9 for now: the
  subscription-manager team is working to make the registration of
  containers easier in newer versions, so this is still needed for now

This hopefully should keep things working as previously expected with no
behaviour change.