Add code quality checks for scripts and Dockerfiles
We continue to add complexity to the uses of this repo. however, we've not had any automated checks
for the sheel scripts or Dockerfiles here. this has made it a bit challenging to review changes that
deviate from the existing conventions.

This change starts with the introduction of shellcheck in the pre-commit config.
We then add an initial test to check Dockerfiles for warnings and errors, ignoring a couple.
This all gets wrapped into a new github action that will be ran for each new pull request.

The largest changes associated with this commit are the updates needed to bring the project in compliance with these new checks.